split app.svelte in components. added a landing page to load audio.
